refactor: 지뢰판 출력하는 로직 개선
